update the /ssj directory to include a [workflow] param in the url, and make updates to set isOperationsGuide in the useUserContext hook
