+++ lib/useUserContext.js
@@ -22,7 +22,7 @@ export function UserProvider({ children }) {
   const [currentUser, setCurrentUser] = useState(null);
   const [loading, setLoading] = useState(true);
   const [isAdmin, setIsAdmin] = useState(false);
-  const [isOperationsGuide, setIsOperationsGuide] = useState(true);
+  const [isOperationsGuide, setIsOperationsGuide] = useState(false);
 
   const router = useRouter();
   useEffect(() => {
@@ -53,6 +53,11 @@ export function UserProvider({ children }) {
             return a.id === personId;
           })?.attributes?.isOnboarded;
 
+          const includedPersonBasic = result.data?.included?.filter(
+            (a) => a.id === personId
+          );
+          // console.log(includedPersonBasic);
+          const opsGuide = includedPersonBasic[0].attributes["isOg?"];
           if (!!userAttributes && !ignore) {
             const user = {
               id: personId,
@@ -64,7 +69,7 @@ export function UserProvider({ children }) {
             };
             setCurrentUser(user);
             setIsAdmin(user?.attributes?.isAdmin);
-            console.log({ result });
+            setIsOperationsGuide(opsGuide);
           }
           console.log("highlight.io identifer", userAttributes.id);
           H.identify(userAttributes.email, {
